requirements.txt
python-binance>=1.0.19
flask>=2.3.3
gunicorn>=21.2.0
sqlalchemy>=2.0.21
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
discord.py>=2.3.2
pandas>=2.0.3
numpy>=1.24.3
ta>=0.10.2
python-dateutil>=2.8.2
pytz>=2023.3
scikit-learn>=1.3.0
matplotlib>=3.7.2
plotly>=5.16.1
scipy>=1.11.2
tensorflow>=2.13.0
keras>=2.13.1
xgboost>=1.7.6
mplfinance>=0.12.7a0
alembic>=1.13.1
joblib>=1.3.2
